fix(InteractiveLogo): move plausible event logging to mouseleave event
Initially I had added a call to log this event to Plausible on the mousemove handler
for the interactive logo.  This had the unwanted result of logging multiple (many)
separate events for each single actual interaction with the logo.

Moving the call to log the event to the mouseleave event ensures that only one event
will ever be logged per actual single interaction.