fix: move plausible hook for theme change
Plausible was not recording theme change events with the hook to Plausible in its
previous location within the actual ThemeContext setTheme() function.  Moving it
to the actual handler for the theme toggle button(s) in the UI, which is more
similar to how Plausible events are being sent with other handlers and events.